// typescript/moonpump/pumpTypes/types/UpdateConfigParams.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as web3 from '@solana/web3.js'
import * as beet from '@metaplex-foundation/beet'
import * as beetSolana from '@metaplex-foundation/beet-solana'
export type UpdateConfigParams = {
  feeRecipient: beet.COption<web3.PublicKey>
  operator: beet.COption<web3.PublicKey>
  tradeFeeBasisPoints: beet.COption<number>
  pumpFee: beet.COption<beet.bignum>
}

/**
 * @category userTypes
 * @category generated
 */
export const updateConfigParamsBeet =
  new beet.FixableBeetArgsStruct<UpdateConfigParams>(
    [
      ['feeRecipient', beet.coption(beetSolana.publicKey)],
      ['operator', beet.coption(beetSolana.publicKey)],
      ['tradeFeeBasisPoints', beet.coption(beet.u16)],
      ['pumpFee', beet.coption(beet.u64)],
    ],
    'UpdateConfigParams'
  )