// typescript/moonpump/pumpTypes/types/PumpParams.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as beet from '@metaplex-foundation/beet'
export type PumpParams = {
  name: string
  ticker: string
  uri: string
}

/**
 * @category userTypes
 * @category generated
 */
export const pumpParamsBeet = new beet.FixableBeetArgsStruct<PumpParams>(
  [
    ['name', beet.utf8String],
    ['ticker', beet.utf8String],
    ['uri', beet.utf8String],
  ],
  'PumpParams'
)