// typescript/moonpump/pumpTypes/types/VanityPumpParams.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as beet from '@metaplex-foundation/beet'
export type VanityPumpParams = {
  seeds: number[] /* size: 32 */
  name: string
  ticker: string
  uri: string
}

/**
 * @category userTypes
 * @category generated
 */
export const vanityPumpParamsBeet =
  new beet.FixableBeetArgsStruct<VanityPumpParams>(
    [
      ['seeds', beet.uniformFixedSizeArray(beet.u8, 32)],
      ['name', beet.utf8String],
      ['ticker', beet.utf8String],
      ['uri', beet.utf8String],
    ],
    'VanityPumpParams'
  )